Refactor and move grant access call|
