+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/CCDCreateCaseService.java
@@ -3,13 +3,20 @@ package uk.gov.hmcts.cmc.claimstore.services.ccd;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
+import uk.gov.hmcts.cmc.claimstore.idam.models.User;
+import uk.gov.hmcts.cmc.claimstore.services.UserService;
 import uk.gov.hmcts.cmc.claimstore.stereotypes.LogExecutionTime;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
+import uk.gov.hmcts.reform.ccd.client.CaseAccessApi;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.EventRequestData;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
+import uk.gov.hmcts.reform.ccd.client.model.UserId;
+
+import static uk.gov.hmcts.cmc.claimstore.repositories.CCDCaseApi.CASE_TYPE_ID;
+import static uk.gov.hmcts.cmc.claimstore.repositories.CCDCaseApi.JURISDICTION_ID;
 
 @Component
 @ConditionalOnProperty(prefix = "feature_toggles", name = "ccd_enabled")
@@ -17,11 +24,19 @@ public class CCDCreateCaseService {
 
     private final CoreCaseDataApi coreCaseDataApi;
     private final AuthTokenGenerator authTokenGenerator;
+    private final CaseAccessApi caseAccessApi;
+    private final UserService userService;
 
     @Autowired
-    public CCDCreateCaseService(CoreCaseDataApi coreCaseDataApi, AuthTokenGenerator authTokenGenerator) {
+    public CCDCreateCaseService(CoreCaseDataApi coreCaseDataApi,
+                                AuthTokenGenerator authTokenGenerator,
+                                CaseAccessApi caseAccessApi,
+                                UserService userService
+    ) {
         this.coreCaseDataApi = coreCaseDataApi;
         this.authTokenGenerator = authTokenGenerator;
+        this.caseAccessApi = caseAccessApi;
+        this.userService = userService;
     }
 
     @LogExecutionTime
@@ -76,6 +91,20 @@ public class CCDCreateCaseService {
             eventRequestData.getCaseTypeId(),
             eventRequestData.isIgnoreWarning(),
             caseDataContent
+        );
+    }
+
+    @LogExecutionTime
+    public void grantAccessToCase(CaseDetails caseDetails, String letterHolderId) {
+        User user = userService.authenticateAnonymousCaseWorker();
+        caseAccessApi.grantAccessToCase(
+            user.getAuthorisation(),
+            authTokenGenerator.generate(),
+            user.getUserDetails().getId(),
+            JURISDICTION_ID,
+            CASE_TYPE_ID,
+            caseDetails.getId().toString(),
+            new UserId(letterHolderId)
         );
     }
