+++ src/test/java/uk/gov/hmcts/cmc/claimstore/services/ccd/CoreCaseDataServiceFailureTest.java
@@ -132,7 +132,6 @@ public class CoreCaseDataServiceFailureTest {
             referenceNumberService,
             coreCaseDataApi,
             authTokenGenerator,
-            caseAccessApi,
             jobSchedulerService,
             ccdCreateCaseService
         );
